frontend(manageCounselor): remvoed the manageCounselor component and moved it to the page
