+++ frontend/src/components/FilterableTable.module.css
@@ -1,15 +1,3 @@
-/* body {
-  font-weight: 400;
-  margin: 0;
-  padding: 0;
-} */
-
-/* .header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-} */
-
 .title {
   font-size: 3rem;
   margin-right: 1.25rem;
@@ -21,12 +9,6 @@
   margin-top: 3vh;
   margin-bottom: 3vh;
 }
-
-/* .search-container {
-  position: relative;
-  width: 25vw;
-  margin-right: 40vw;
-} */
 
 .search-bar {
   width: 30rem;
